+++ tests/geometries/polygon_test.py
@@ -20,6 +20,7 @@ import pygeoif.geometry as geo
 
 from fastkml.enums import AltitudeMode
 from fastkml.enums import Verbosity
+from fastkml.geometry import OuterBoundaryIs
 from fastkml.geometry import Polygon
 from tests.base import Lxml
 from tests.base import StdLibrary
@@ -244,7 +245,9 @@ class TestStdLibrary(StdLibrary):
         polygon = Polygon.class_from_string(doc)
 
         assert not polygon.geometry
-        assert polygon.outer_boundary_is is not None
+        assert polygon.outer_boundary is not None
+        assert isinstance(polygon.outer_boundary, OuterBoundaryIs)
+        assert len(polygon.inner_boundaries) == 0
         assert "tessellate>1</" in polygon.to_string()