+++ tests/kml_test.py
@@ -20,6 +20,7 @@ import pathlib
 
 import pygeoif as geo
 
+from fastkml import containers
 from fastkml import features
 from fastkml import kml
 from fastkml.containers import Document
@@ -57,6 +58,49 @@ class TestStdLibrary(StdLibrary):
         )
         k2 = kml.KML.class_from_string(k.to_string(), ns="")
         assert k.to_string() == k2.to_string()
+
+    def test_kml_with_document(self) -> None:
+        """Kml file with document/folder/placemark/polygon."""
+        doc = (
+            '<kml xmlns="http://www.opengis.net/kml/2.2">'
+            '<Document id="root_doc">'
+            "<Folder><name>test</name>"
+            "<Placemark>"
+            "<Polygon>"
+            "<outerBoundaryIs>"
+            "<LinearRing><coordinates>"
+            "-93.7426720226024,57.4519411370713 -93.6051809086549,49.4316261567984 "
+            "-80.8643376828499,49.5232868994301 -81.2309806533767,57.4519411370713 "
+            "-81.2309806533767,57.4519411370713 -93.7426720226024,57.4519411370713"
+            "</coordinates></LinearRing>"
+            "</outerBoundaryIs>"
+            "<innerBoundaryIs>"
+            "<LinearRing><coordinates>"
+            "-91.8663227028478,56.050879726904 -91.7704563496718,53.9897531336206 "
+            "-90.1407283456804,54.0856194867966 -90.0927951690924,56.002946550316 "
+            "-91.8663227028478,56.050879726904"
+            "</coordinates></LinearRing>"
+            "</innerBoundaryIs>"
+            "<innerBoundaryIs>"
+            "<LinearRing><coordinates>"
+            "-85.4912102166459,55.90708019714 -85.4912102166459,54.0376863102086 "
+            "-83.8135490360665,54.0856194867966 -83.9094153892425,55.90708019714 "
+            "-85.4912102166459,55.90708019714</coordinates></LinearRing>"
+            "</innerBoundaryIs>"
+            "</Polygon>"
+            "</Placemark>"
+            "</Folder>"
+            "</Document></kml>"
+        )
+        k = kml.KML.class_from_string(doc)
+        assert len(k.features) == 1
+        assert isinstance(k.features[0], Document)
+        assert len(k.features[0].features) == 1
+        assert isinstance(k.features[0].features[0], containers.Folder)
+        assert len(k.features[0].features[0].features) == 1
+        assert isinstance(k.features[0].features[0].features[0], features.Placemark)
+        assert isinstance(k.features[0].features[0].features[0].geometry, geo.Polygon)
+        assert len(list(k.features[0].features[0].features[0].geometry.interiors)) == 2
 
 
 class TestParseKML(StdLibrary):